Refactor project structure and enhance UI components
- Removed Navigation component from MITDataSciencePage, NotFoundPage, and PersonalProjectsPage for a cleaner layout.
- Introduced a new Layout component to encapsulate Navigation and ConnectLinks for consistent page structure.
- Created ProjectCard component to streamline project display logic and improve code reusability.
- Updated ProjectDetails and ProjectDetails.module.css to reflect new styles and structure.
- Consolidated project data into separate files for better organization and maintainability.
- Enhanced styling for tech tags and section titles in ProjectDetails.module.css for improved visual hierarchy.
- Added new data structures for books and personal projects to facilitate future expansion.
- Removed unused styles from ProjectsPage.module.css to optimize CSS file size.
