+++ frontend/src/pages/AboutPage/AboutPage.module.css
@@ -167,17 +167,16 @@
 
 .timeline {
   max-width: 820px;
-  margin: 0;
   position: relative;
-  padding-left: 2.5rem;
+  padding-left: 2rem;
 }
 
-/* Vertical line */
+/* Vertical line — starts at center of first dot, not above it */
 .timeline::before {
   content: '';
   position: absolute;
-  left: 0.75rem;
-  top: 0;
+  left: 0;
+  top: 0.8rem;   /* dot top (0.45rem) + half dot height (0.35rem) */
   bottom: 0;
   width: 1px;
   background: var(--color-rule);
@@ -185,28 +184,25 @@
 
 .timeline-item {
   position: relative;
-  margin-bottom: 2.5rem;
+  padding-bottom: 2.5rem;
 }
 
 .timeline-item:last-child {
-  margin-bottom: 0;
+  padding-bottom: 0;
 }
 
+/* Dot — left: 0 on ::before == outer-left of .timeline
+   .timeline-item content starts 2rem right of that
+   center dot (0.7rem wide) on the line: -(2rem + 0.35rem) */
 .timeline-dot {
   position: absolute;
-  left: -1.95rem;
-  top: 0.6rem;
-  width: 0.6rem;
-  height: 0.6rem;
+  left: calc(-2rem - 0.35rem);
+  top: 0.45rem;
+  width: 0.7rem;
+  height: 0.7rem;
   border-radius: 50%;
   background: var(--color-primary);
-  border: 2px solid var(--color-bg-subtle);
-}
-
-.timeline-content {
-  padding: 0 0 0.5rem;
-  position: relative;
-  transition: none;
+  outline: 3px solid var(--color-bg-subtle);
 }
 
 .timeline-content h3 {
@@ -255,8 +251,6 @@
   display: grid;
   grid-template-columns: repeat(2, 1fr);
   gap: 2rem;
-  max-width: 960px;
-  margin: 0 auto;
 }
 
 .capability-card {
@@ -322,8 +316,6 @@
 }
 
 .education-grid {
-  max-width: 780px;
-  margin: 0 auto;
   display: grid;
   gap: 2rem;
 }
@@ -376,8 +368,7 @@
 }
 
 .awards-grid {
-  max-width: 780px;
-  margin: 0 auto 3rem;
+  margin: 0 0 3rem;
 }
 
 .award-card {
@@ -426,7 +417,6 @@
   display: grid;
   grid-template-columns: repeat(2, 1fr);
   gap: 0.5rem;
-  max-width: 820px;
 }
 
 .cert-card {
@@ -464,14 +454,6 @@
     font-size: 1rem;
   }
 
-  .timeline {
-    padding-left: 2rem;
-  }
-
-  .timeline-dot {
-    left: -1.35rem;
-  }
-
   .capabilities-grid {
     grid-template-columns: 1fr;
   }
@@ -497,10 +479,6 @@
   .hero-photo {
     width: 120px;
     height: 120px;
-  }
-
-  .timeline-content {
-    padding: 1.25rem;
   }
 
   .capability-card,
