+++ frontend/src/pages/HomePage/HomePage.module.css
@@ -434,75 +434,6 @@
   margin-top: 0.5rem;
 }
 
-.featuredCard {
-  background: var(--color-bg-elevated);
-  padding: 1.75rem;
-  border-radius: var(--border-radius-lg);
-  border: 1px solid var(--color-border);
-  display: flex;
-  flex-direction: column;
-  transition: border-color 0.2s ease, transform 0.2s ease;
-}
-
-.featuredCard:hover {
-  border-color: var(--color-primary);
-  transform: translateY(-2px);
-}
-
-.featuredBadges {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.5rem;
-  margin-bottom: 1rem;
-}
-
-.featuredBadge {
-  background: transparent;
-  color: var(--color-text-muted);
-  padding: 0.2rem 0.55rem;
-  border: 1px solid var(--color-border);
-  border-radius: 3px;
-  font-family: var(--font-mono);
-  font-size: 0.7rem;
-  font-weight: 500;
-  letter-spacing: 0.05em;
-  text-transform: uppercase;
-}
-
-.featuredCard h3 {
-  font-family: var(--font-display);
-  color: var(--color-text-primary);
-  font-size: 1.15rem;
-  font-weight: 600;
-  margin-bottom: 0.5rem;
-  letter-spacing: -0.015em;
-  line-height: 1.25;
-}
-
-.featuredCard p {
-  color: var(--color-text-secondary);
-  font-size: 0.9rem;
-  line-height: 1.6;
-  flex: 1;
-}
-
-.featuredLink {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.4rem;
-  margin-top: 1.25rem;
-  color: var(--color-primary);
-  font-family: var(--font-mono);
-  font-weight: 500;
-  font-size: 0.8rem;
-  letter-spacing: 0.05em;
-  text-transform: uppercase;
-  transition: gap 0.2s ease;
-}
-
-.featuredLink:hover {
-  gap: 0.7rem;
-}
 
 .featuredFooter {
   text-align: left;
