+++ frontend/src/pages/PersonalProjectsPage/PersonalProjectsPage.tsx
@@ -1,42 +1,7 @@
 import React from 'react';
-import Navigation from '../../components/Navigation/Navigation';
 import Seo from '../../components/Seo';
+import { PROJECTS } from '../../data/personalProjects';
 import styles from './PersonalProjectsPage.module.css';
-
-interface PersonalProject {
-  title: string;
-  summary: string;
-  tech: string[];
-  href: string;
-  youtube?: string;
-}
-
-const PROJECTS: PersonalProject[] = [
-  {
-    title: 'Facial Detection System',
-    summary:
-      'Real-time face detection and emotion recognition using deep-learning models and OpenCV for live video processing.',
-    tech: ['OpenCV', 'Python', 'Deep Learning', 'Real-time Processing'],
-    href: 'https://github.com/ianedmundson1/Facial-detection',
-    youtube: 'https://www.youtube.com/watch?v=0VEvEf_r25U'
-  },
-  {
-    title: 'Lane Detection Algorithm',
-    summary:
-      'Lane detection for autonomous-driving applications using computer vision and image-processing techniques, robust across varying lighting and road conditions.',
-    tech: ['Computer Vision', 'Image Processing', 'OpenCV', 'Autonomous Vehicles'],
-    href: 'https://github.com/ianedmundson1/Lane-detection',
-    youtube: 'https://www.youtube.com/watch?v=0klrGBsJtYY'
-  },
-  {
-    title: 'IoT Security Camera System',
-    summary:
-      'Raspberry Pi-based security camera with motion detection, cloud storage, and automated notifications for end-to-end home surveillance.',
-    tech: ['Raspberry Pi', 'IoT', 'Motion Detection', 'Cloud Storage'],
-    href: 'https://github.com/ianedmundson1/Security-camera',
-
-  },
-];
 
 const PersonalProjectsPage: React.FC = () => {
   return (
@@ -46,7 +11,6 @@ const PersonalProjectsPage: React.FC = () => {
         description="Side projects by Ian Edmundson — computer vision with OpenCV, IoT systems on Raspberry Pi, and deep-learning experiments built outside of work."
         path="/projects/personal"
       />
-      <Navigation />
       <main className="main-content">
         <section className={styles.hero}>
           <div className={styles.heroContainer}>
