+++ frontend/src/pages/ProjectsPage/ProjectsPage.module.css
@@ -146,98 +146,6 @@
   margin-bottom: 2rem;
 }
 
-.project-card {
-  background: var(--color-bg-elevated);
-  border-radius: var(--border-radius-lg);
-  padding: 1.75rem;
-  border: 1px solid var(--color-border);
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  transition: border-color 0.2s ease, transform 0.2s ease;
-}
-
-.project-card:hover {
-  border-color: var(--color-primary);
-  transform: translateY(-2px);
-}
-
-.project-header {
-  margin-bottom: 1rem;
-}
-
-.project-header h3 {
-  font-family: var(--font-display);
-  font-size: 1.2rem;
-  color: var(--color-text-primary);
-  margin-bottom: 0.75rem;
-  font-weight: 600;
-  letter-spacing: -0.015em;
-  line-height: 1.25;
-}
-
-.project-badges {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.5rem;
-}
-
-.project-badge {
-  background: transparent;
-  color: var(--color-text-muted);
-  padding: 0.25rem 0.6rem;
-  border-radius: 3px;
-  font-family: var(--font-mono);
-  font-size: 0.7rem;
-  font-weight: 500;
-  letter-spacing: 0.05em;
-  text-transform: uppercase;
-  border: 1px solid var(--color-border);
-}
-
-.project-badge.interactive {
-  background: var(--color-primary);
-  color: var(--color-bg);
-  border-color: var(--color-primary);
-}
-
-.project-description {
-  color: var(--color-text-secondary);
-  line-height: 1.7;
-  margin-bottom: 1.5rem;
-  font-size: 0.95rem;
-}
-
-.project-links {
-  display: flex;
-  gap: 0.75rem;
-}
-
-.project-link {
-  text-decoration: none;
-  display: inline-flex;
-  align-items: center;
-  gap: 0.4rem;
-  padding: 0.5rem 1rem;
-  border-radius: var(--border-radius);
-  font-family: var(--font-body);
-  font-weight: 500;
-  font-size: 0.875rem;
-  transition: all 0.2s ease;
-  border: 1px solid transparent;
-}
-
-.project-link.primary {
-  background: var(--color-text-primary);
-  color: var(--color-bg);
-  border-color: var(--color-text-primary);
-}
-
-.project-link.primary:hover {
-  background: var(--color-primary);
-  border-color: var(--color-primary);
-}
-
 /* ==========================================
    CATEGORY FOOTER
    ========================================== */
